requirements.txt
bittensor
boto3
safetensors
torch
transformers
python-dotenv
datasets
torchvision
wandb
typer